Added event handlers to trainer hparams registry, test fixes, moved JSONTraceHandler import to top
